+++ src/Client.php
@@ -232,9 +232,10 @@ final class Client implements ClientInterface
         $event->setTags($this->options->getTags());
         $event->setEnvironment($this->options->getEnvironment());
 
+        $isTransaction = EventType::transaction() === $event->getType();
         $sampleRate = $this->options->getSampleRate();
 
-        if (EventType::transaction() !== $event->getType() && $sampleRate < 1 && mt_rand(1, 100) / 100.0 > $sampleRate) {
+        if (!$isTransaction && $sampleRate < 1 && mt_rand(1, 100) / 100.0 > $sampleRate) {
             $this->logger->info('The event will be discarded because it has been sampled.', ['event' => $event]);
 
             return null;
@@ -251,11 +252,13 @@ final class Client implements ClientInterface
             }
         }
 
-        $previousEvent = $event;
-        $event = ($this->options->getBeforeSendCallback())($event);
+        if (!$isTransaction) {
+            $previousEvent = $event;
+            $event = ($this->options->getBeforeSendCallback())($event);
 
-        if (null === $event) {
-            $this->logger->info('The event will be discarded because the "before_send" callback returned "null".', ['event' => $previousEvent]);
+            if (null === $event) {
+                $this->logger->info('The event will be discarded because the "before_send" callback returned "null".', ['event' => $previousEvent]);
+            }
         }
 
         return $event;